various improvements & bug fixes

+++ engine/include/scene_manager.hpp
@@ -6,17 +6,17 @@
 namespace kuki {
 class KUKI_ENGINE_API SceneManager {
 private:
-  unsigned int nextId{0};
-  std::unordered_map<unsigned int, Scene*> idToScene;
-  std::unordered_map<std::string, unsigned int> nameToId;
+  size_t nextId{0};
+  std::unordered_map<size_t, Scene*> idToScene;
+  std::unordered_map<std::string, size_t> nameToId;
 public:
   ~SceneManager();
-  unsigned int Create(const std::string&);
-  void Delete(unsigned int);
+  size_t Create(const std::string&);
+  void Delete(size_t);
   void Delete(const std::string&);
-  Scene* Get(unsigned int);
+  Scene* Get(size_t);
   Scene* Get(const std::string&);
-  bool Has(unsigned int);
+  bool Has(size_t);
   bool Has(const std::string&);
 };
 } // namespace kuki
